refactor: 중복되는 template코드를 재사용 가능한 모듈로 분리
