+++ src/js/utils/localStorage.js
@@ -2,7 +2,7 @@ import { ERROR_MESSAGE } from './constant.js';
 
 export default {
   set: (key, value) => {
-    if (typeof value !== 'string' || !Array.isArray(value)) {
+    if (typeof value !== 'string' && !Array.isArray(value)) {
       throw new Error(
         ERROR_MESSAGE.ONLY_STRING_OR_ARRAY_CAN_BE_STORED_IN_LOCAL_STORAGE,
       );
