+++ src/js/view/main.js
@@ -1,53 +1,10 @@
 import { $ } from '../utils/querySelector.js';
-
-const savedClipTemplate = (video) => {
-  const [year, month, day] = new Date(video.snippet.publishTime)
-    .toLocaleDateString()
-    .split('.');
-
-  return `
-  <article class="clip">
-    <div class="preview-container">
-      <iframe 
-        width="100%"
-        height="118"
-        src=https://www.youtube.com/embed/${video.id.videoId}
-        frameborder="0"
-        allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture"
-        allowfullscreen
-      ></iframe>
-    </div>
-    <div class="content-container pt-2 px-1">
-      <h3>${video.snippet.title}</h3>
-      <div>
-        <a
-          href="https://www.youtube.com/channel/UC-mOekGSesms0agFntnQang"
-          target="_blank"
-          class="channel-name mt-1"
-        >
-          ${video.snippet.channelTitle}
-        </a>
-        <div class="meta">
-          <p>${year}년 ${month}월 ${day}일</p>
-        </div>
-        <div class="d-flex justify-end">
-          <div>
-            <span class="opacity-hover">✅</span>
-            <span class="opacity-hover">👍</span>
-            <span class="opacity-hover">💬</span>
-            <span class="opacity-hover">🗑️</span>
-          </div>
-        </div>
-      </div>
-    </div>
-  </article>
-`;
-};
+import { clipTemplate } from './clipTemplate.js';
 
 export const renderSavedClips = (savedClips) => {
   $(
     '[data-js=saved-page__video-wrapper]',
   ).innerHTML = savedClips
-    .map((savedClip, index) => savedClipTemplate(savedClip, index))
+    .map((savedClip, index) => clipTemplate(savedClip, index, {}))
     .join('');
 };
