+++ src/js/view/modal.js
@@ -1,52 +1,7 @@
 import { $ } from '../utils/querySelector.js';
-import date from '../utils/date.js';
+import { clipTemplate } from './clipTemplate.js';
 
 const $modal = $('[data-js="youtube-search-modal"]');
-
-const YMDtemplate = (time) => {
-  const [year, month, day] = date.toYMDArray(time);
-
-  return `<p>${year}년 ${month}월 ${day}일</p>`;
-};
-
-const searchResultClipTemplate = (video, index, isSaved = false) => {
-  return `
-    <article class="clip" data-js="youtube-search-modal__clip">
-      <div class="preview-container">
-        <iframe 
-          width="100%"
-          height="118"
-          src=https://www.youtube.com/embed/${video.id.videoId}
-          frameborder="0"
-          allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture"
-          allowfullscreen
-        ></iframe>
-      </div>
-      <div class="content-container pt-2 px-1">
-        <h3>${video.snippet.title}</h3>
-        <div>
-          <a
-            href="https://www.youtube.com/channel/UC-mOekGSesms0agFntnQang"
-            target="_blank"
-            class="channel-name mt-1"
-          >
-            ${video.snippet.channelTitle}
-          </a>
-          <div class="meta">
-            ${YMDtemplate(video.snippet.publishTime)}
-          </div>
-          <div class="d-flex justify-end">
-          ${
-            isSaved
-              ? ''
-              : `<button class="btn" data-js="save-button" data-clip-index=${index}>⬇️ 저장</button>`
-          }      
-          </div>
-        </div>
-      </div>
-    </article>
-  `;
-};
 
 export const openModal = () => {
   $modal.classList.add('open');
@@ -79,7 +34,7 @@ export const renderExtraClips = (videoItems, savedClipIds) => {
   $('[data-js=youtube-search-modal__video-wrapper]').innerHTML += videoItems
     .map((video, index) => {
       const isSaved = savedClipIds.includes(video.id.videoId);
-      return searchResultClipTemplate(video, index, isSaved);
+      return clipTemplate(video, index, { isModal: true, isSaved });
     })
     .join('');
 };
@@ -88,7 +43,7 @@ export const renderClips = (videoItems, savedClipIds) => {
   $('[data-js=youtube-search-modal__video-wrapper]').innerHTML = videoItems
     .map((video, index) => {
       const isSaved = savedClipIds.includes(video.id.videoId);
-      return searchResultClipTemplate(video, index, isSaved);
+      return clipTemplate(video, index, { isModal: true, isSaved });
     })
     .join('');
 };
